+++ content/test/integration/logic/snapshot-cleaner.spec.ts
@@ -4,14 +4,14 @@ import { mkdirSync, mkdtempSync, rmSync } from 'fs'
 import os from 'os'
 import path from 'path'
 import { promisify } from 'util'
-import { EnvironmentConfig } from '../../src/Environment'
-import { bufferToStream } from '../../src/ports/contentStorage/contentStorage'
-import { ServiceImpl } from '../../src/service/ServiceImpl'
-import { cleanSnapshots } from '../../src/snapshotCleaner'
-import { makeNoopValidator } from '../helpers/service/validations/NoOpValidator'
-import { loadStandaloneTestEnvironment, testCaseWithComponents } from './E2ETestEnvironment'
-import { buildDeployData } from './E2ETestUtils'
-import { getIntegrationResourcePathFor } from './resources/get-resources-path'
+import { EnvironmentConfig } from '../../../src/Environment'
+import { cleanSnapshots } from '../../../src/logic/snapshot-cleaner'
+import { bufferToStream } from '../../../src/ports/contentStorage/contentStorage'
+import { ServiceImpl } from '../../../src/service/ServiceImpl'
+import { makeNoopValidator } from '../../helpers/service/validations/NoOpValidator'
+import { loadStandaloneTestEnvironment, testCaseWithComponents } from '../E2ETestEnvironment'
+import { buildDeployData } from '../E2ETestUtils'
+import { getIntegrationResourcePathFor } from '../resources/get-resources-path'
 
 const promifiedExec = promisify(exec)
